usize in a hash preimage: two live sites where a browser and a native host committed to different bytes from identical inputs
With `wasm` compiling again, the width class is measurable for the first time since the finding that
named it (`hash_index` feeding `usize::to_le_bytes()` into a BLAKE3 preimage: 4 bytes on wasm32, 8 on
x86_64). Swept the 67 local crates in the wasm32 graph — 3594 `to_*_bytes()` call sites, 31 whose
receiver could be a `usize`, and after reading all 31, **two are real**. The other 29 are the correct
pattern already (`u32::try_from(x.len())`, `descriptor_ir2_canonical::Writer::index`'s explicit
`u64::try_from`, or plainly `u32`/`u64`-typed) — that is worth saying, because it means the discipline
mostly holds and these two are outliers, not a tide.

## MEASURED, not read — the REAL functions, driven on both targets

A probe crate linked `dregg-intent` + `dregg-bridge` and was built twice, once for
`aarch64-apple-darwin` and once for `wasm32-unknown-unknown` (driven in node). Identical fixed inputs
— fixed ciphertext row, fixed 32-byte session secret `0xa5…`, fixed nonce `0x5a…`, fixed key `0x11…`:

| call | aarch64 BEFORE | wasm32 BEFORE | aarch64 AFTER | wasm32 AFTER |
|---|---|---|---|---|
| `pir_row_probe(0)` | 3833918878 | **1093993295** | 3833918878 | **3833918878** |
| `pir_row_probe(1)` | 430504304 | **4214063577** | 430504304 | **430504304** |
| `pir_row_probe(3)` | 2395856935 | **852431118** | 2395856935 | **2395856935** |
| `hash_index(0,0)` | 1669440319 | **1558419647** | 1669440319 | **1669440319** |
| `hash_index(1,2)` | 1209911992 | **1651138428** | 1209911992 | **1209911992** |
| `hash_index(7,3)` | 568021628 | **142495896** | 568021628 | **568021628** |

⚑ Every native column is UNCHANGED by the fix. `usize as u64` is byte-identical on a 64-bit host, so
nothing native re-emits, no descriptor moves, no VK rotates, and no schema epoch bumps. The browser
converges onto what the native side already computed.

## The two sites

### `intent/src/pir.rs` — download-all PIR, both halves

`EncryptedDatabase::encrypt` (the `.enumerate()` binder) and `decrypt_row(row_idx: usize, …)` each
build the per-row key as
`blake3::derive_key("dregg-pir-download-all-row-key", secret ++ row_idx.to_le_bytes() ++ nonce)`.
A native server and a browser client derived DIFFERENT keys for the same row. The row cipher is a
bare XOR keystream with **no authentication tag**, so the tab did not get an error — it got silently
wrong `BabyBear` felts and decoded them as a row. ⚠ Row 0 diverged too: what differs is the concat
LENGTH, not the digits, so this was never a large-index edge case. `dregg-intent` is a direct
dependency of `dregg-wasm`.

### `bridge/src/present.rs::hash_index` — the exact function the original finding named

`hash_index(level: usize, sibling_idx: usize, key)` folds a BLAKE3 digest to a `BabyBear` sibling
felt, so the two targets built different Merkle trees. It is `pub` and unconditionally compiled into
the wasm32 bundle; its callers are all `#[cfg(any(test, feature = "test-utils"))]`, which bounds the
blast radius but does not make a `pub` function in the shipped bundle right.

`tests/src/wire_format_e2e.rs` carries a hand COPY of it (that module is unconditional in
`tests/src/main.rs`, while `dregg-bridge` is an OPTIONAL dep of that crate gated to
`__legacy_tests`/`__wip_tests` — the copy exists to avoid the dep). Fixed identically and now
labelled as a copy naming its original. ⚠ NAMED RESIDUAL, not closed: it is still a twin, and the
un-twinning is a decision about that crate's feature graph — the gate's comment says bridge is kept
optional "so the workspace stays buildable while bridge undergoes token-vocabulary changes", which is
someone's live constraint, not mine to silently drop.

## THE GATE — `wasm/tests/cross_target_preimage_widths.rs`, and why it runs twice

A fix here is invisible to a native test: on a 64-bit host `usize` and `u64` are the same bytes, so a
host-only test cannot go red on a revert. The gate is therefore a PAIR of legs over the same
constants, and every expectation RESTATES the preimage with an explicit `[u8; 8]` written digit by
digit — it never calls `usize::to_le_bytes` itself, so it means the same thing on both targets.

    cargo test -p dregg-wasm --test cross_target_preimage_widths    2 passed
    wasm-pack test --node wasm --test cross_target_preimage_widths  3 passed

The wasm32 leg carries `usize_is_four_bytes_on_this_target`, so it cannot silently become
meaningless, and `hash_index_…` carries a DISCRIMINATOR: an `assert_ne!` against the 4-byte-index
felt, which on wasm32 is exactly the pre-fix value (1651138428 for `(1,2)`). Reverting either `as
u64` breaks the wasm32 leg and leaves the host leg green — which is the honest shape of this class.

`dregg-bridge` becomes a `[dev-dependencies]` entry of `dregg-wasm` for the pin. It is already in
this graph transitively via `dregg-intent`, so it adds no compilation; the shipped cdylib does not
name it.

## Verified

* `cargo check -p dregg-intent -p dregg-bridge --all-targets` — 0 errors.
* `cargo check -p dregg-tests --all-targets` — 0 errors.
* `cargo check --target wasm32-unknown-unknown --all-targets` (wasm workspace) — 0 errors.
* The before/after probe table above, both targets, real functions.

+++ bridge/src/present.rs
@@ -1904,8 +1904,14 @@ pub fn verify_revealed_facts_commitment(
 /// MUST NOT be used in production.
 pub fn hash_index(level: usize, sibling_idx: usize, key: &[u8; 32]) -> u32 {
     let mut hasher = blake3::Hasher::new();
-    hasher.update(&level.to_le_bytes());
-    hasher.update(&sibling_idx.to_le_bytes());
+    // ⚑ `as u64`, NOT a bare `usize`. This crate is in the wasm32 bundle's graph, and
+    // `usize::to_le_bytes()` is 8 bytes on x86_64/aarch64 and 4 on wasm32 — so a native
+    // prover and a browser light client built provably different sibling felts, and
+    // therefore different Merkle trees, from identical inputs with nothing red. Measured
+    // 2026-08-01: `hash_index(1, 2, [0x11; 32])` = 1209911992 native, 1651138428 on wasm32.
+    // On a 64-bit host `usize as u64` is byte-identical, so no native felt changes.
+    hasher.update(&(level as u64).to_le_bytes());
+    hasher.update(&(sibling_idx as u64).to_le_bytes());
     hasher.update(key);
     let hash = hasher.finalize();
     let bytes = hash.as_bytes();
